+++ src/changelog.md
@@ -1,3 +1,7 @@
+# Release 3.0.0
+* Requires AzureStack Update 2203+
+## Azs.Compute.Admin module
+* Get-AzsComputeScaleUnit - Added admin support to view the tenant VM and compute node information of a target scale unit
 
 # Release 2.2.0
 * Requires AzureStack Update 2108+